move Operations.sameLanguage/subsetOf to AutomatonTestUtil in test-framework
