+++ lucene/core/src/test/org/apache/lucene/util/automaton/TestMinimize.java
@@ -28,7 +28,7 @@ public class TestMinimize extends LuceneTestCase {
       Automaton a = AutomatonTestUtil.randomAutomaton(random());
       Automaton la = Operations.determinize(Operations.removeDeadStates(a), Integer.MAX_VALUE);
       Automaton lb = MinimizationOperations.minimize(a, Integer.MAX_VALUE);
-      assertTrue(Operations.sameLanguage(la, lb));
+      assertTrue(AutomatonTestUtil.sameLanguage(la, lb));
     }
   }
 
@@ -42,7 +42,7 @@ public class TestMinimize extends LuceneTestCase {
       Automaton a = AutomatonTestUtil.randomAutomaton(random());
       a = AutomatonTestUtil.minimizeSimple(a);
       Automaton b = MinimizationOperations.minimize(a, Integer.MAX_VALUE);
-      assertTrue(Operations.sameLanguage(a, b));
+      assertTrue(AutomatonTestUtil.sameLanguage(a, b));
       assertEquals(a.getNumStates(), b.getNumStates());
       int numStates = a.getNumStates();
